+++ modules/terraform-zscc-identity-azure/main.tf
@@ -1,4 +1,7 @@
-# Reference inputs to obtain an existing User Managed Identity Resource to associate to Cloud Connector VM
+################################################################################
+# Reference inputs to obtain an existing User Managed Identity Resource 
+# to associate to Cloud Connector VM
+################################################################################
 data "azurerm_user_assigned_identity" "selected" {
   name                = var.cc_vm_managed_identity_name
   resource_group_name = var.cc_vm_managed_identity_rg