+++ modules/terraform-zscc-workload-azure/main.tf
@@ -1,6 +1,9 @@
-resource "azurerm_network_security_group" "server-nsg" {
+################################################################################
+# Create NSG and Rules for worload VMs
+################################################################################
+resource "azurerm_network_security_group" "workload-nsg" {
   count               = var.workload_count
-  name                = "${var.name_prefix}-server-${count.index + 1}-nsg-${var.resource_tag}"
+  name                = "${var.name_prefix}-workload-${count.index + 1}-nsg-${var.resource_tag}"
   location            = var.location
   resource_group_name = var.resource_group
 
@@ -43,14 +46,18 @@ resource "azurerm_network_security_group" "server-nsg" {
   tags = var.global_tags
 }
 
-resource "azurerm_network_interface" "server-nic" {
+
+################################################################################
+# Create Network Interface and association NSG
+################################################################################
+resource "azurerm_network_interface" "workload-nic" {
   count               = var.workload_count
-  name                = "${var.name_prefix}-server-${count.index + 1}-nic-${var.resource_tag}"
+  name                = "${var.name_prefix}-workload-${count.index + 1}-nic-${var.resource_tag}"
   location            = var.location
   resource_group_name = var.resource_group
 
   ip_configuration {
-    name                          = "${var.name_prefix}-server-${count.index + 1}-nic-conf-${var.resource_tag}"
+    name                          = "${var.name_prefix}-workload-${count.index + 1}-nic-conf-${var.resource_tag}"
     subnet_id                     = var.subnet_id
     private_ip_address_allocation = "dynamic"
   }
@@ -60,22 +67,26 @@ resource "azurerm_network_interface" "server-nic" {
   tags = var.global_tags
 }
 
-resource "azurerm_network_interface_security_group_association" "server-nic-association" {
+resource "azurerm_network_interface_security_group_association" "workload-nic-association" {
   count                     = var.workload_count
-  network_interface_id      = azurerm_network_interface.server-nic[count.index].id
-  network_security_group_id = azurerm_network_security_group.server-nsg[count.index].id
+  network_interface_id      = azurerm_network_interface.workload-nic[count.index].id
+  network_security_group_id = azurerm_network_security_group.workload-nsg[count.index].id
 }
 
-resource "azurerm_linux_virtual_machine" "server-vm" {
+
+################################################################################
+# Create Workload VMs
+################################################################################
+resource "azurerm_linux_virtual_machine" "workload-vm" {
   count               = var.workload_count
-  name                = "${var.name_prefix}-server-vm-${count.index + 1}-${var.resource_tag}"
+  name                = "${var.name_prefix}-workload-vm-${count.index + 1}-${var.resource_tag}"
   location            = var.location
   resource_group_name = var.resource_group
 
-  network_interface_ids = [azurerm_network_interface.server-nic[count.index].id]
+  network_interface_ids = [azurerm_network_interface.workload-nic[count.index].id]
   size                  = var.instance_size
   admin_username        = var.server_admin_username
-  computer_name         = "${var.name_prefix}-server-${count.index + 1}-${var.resource_tag}"
+  computer_name         = "${var.name_prefix}-workload-${count.index + 1}-${var.resource_tag}"
   admin_ssh_key {
     username   = var.server_admin_username
     public_key = "${trimspace(var.ssh_key)} ${var.server_admin_username}@me.io"
@@ -96,7 +107,7 @@ resource "azurerm_linux_virtual_machine" "server-vm" {
   tags = var.global_tags
 
   depends_on = [
-    azurerm_network_interface.server-nic,
-    azurerm_network_interface_security_group_association.server-nic-association
+    azurerm_network_interface.workload-nic,
+    azurerm_network_interface_security_group_association.workload-nic-association
   ]
 }