+++ modules/terraform-zscc-workload-azure/output.tf
@@ -1,3 +1,3 @@
 output "private_ip" {
-  value = azurerm_network_interface.server-nic.*.private_ip_address
+  value = azurerm_network_interface.workload-nic.*.private_ip_address
 }