+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 * added TF_DATA_DIR to zsec and backend.tf to each deployment type to maintain root path as "examples/" directory for all deployments
 * dropped "custom" from deployment types as brownfield is implicit
 * vm_count replaced with workload_count for clarity
+* network infrastructure resources consolidated to terraform-zscc-network-azure module. Inputs refactored.